update directory & imports
