+++ src/index.tsx
@@ -1,9 +1,9 @@
 import App from './App'
 import React from 'react'
-import ReactDOM from 'react-dom'
+import { render } from 'react-dom'
 // import reportWebVitals from './reportWebVitals'
 
-ReactDOM.render(
+render(
   <React.StrictMode>
     <App />
   </React.StrictMode>,
