+++ src/pages/teams/TeamsPage.tsx
@@ -1,20 +1,28 @@
-import '../TeamPage.css'
+import './TeamsPage.css'
 
 import Col from 'react-bootstrap/Col'
 import Container from 'react-bootstrap/Container'
-import MainCard from '../../components/Cards/MainCard/MainCard'
-import Image from '../../assets/img/placeholder.jpeg'
-import MembersCard from '../../components/Cards/MembersCard/MembersCard'
+import MainCard from '@components/Cards/MainCard/MainCard'
+import Image from '@assets/img/placeholder.jpeg'
+import MembersCard from '@components/Cards/MembersCard/MembersCard'
 import React from 'react'
 import Row from 'react-bootstrap/Row'
 
-export const ProjectTeamPage = () => {
-  const card = {
-    team: 'ProjectTeamPage'
+// interface member {
+//   firstName: string
+//   lastName?: string
+//   photo_link?: string = placeholder.jpg
+// }
+
+// interface leader {extend member}
+
+export const TeamPage: React.FC = () => {
+  const card: { team: string } = {
+    team: 'TeamPage'
   }
 
   const members = {
-    teamName: 'Project Team',
+    teamName: '__ Team',
     leader: { name: 'FirstName LastName' },
     member: [
       { name: 'FirstName LastName' },
@@ -63,4 +71,4 @@ export const ProjectTeamPage = () => {
   )
 }
 
-export default ProjectTeamPage
+export default TeamPage
