+++ src/pages/teams/team/Event.tsx
@@ -1,20 +1,20 @@
-import '../TeamPage.css'
+import '../TeamsPage.css'
 
 import Col from 'react-bootstrap/Col'
 import Container from 'react-bootstrap/Container'
-import MainCard from '../../components/Cards/MainCard/MainCard'
-import Image from '../../assets/img/placeholder.jpeg'
-import MembersCard from '../../components/Cards/MembersCard/MembersCard'
+import MainCard from '@components/Cards/MainCard/MainCard'
+import Image from '@assets/img/placeholder.jpeg'
+import MembersCard from '@components/Cards/MembersCard/MembersCard'
 import React from 'react'
 import Row from 'react-bootstrap/Row'
 
-export const MarketingTeamPage = () => {
+export const EventTeamPage = () => {
   const card = {
-    team: 'MarketingTeamPage'
+    team: 'EventTeamPage'
   }
 
   const members = {
-    teamName: 'Marketing Team',
+    teamName: 'Event Team',
     leader: { name: 'FirstName LastName' },
     member: [
       { name: 'FirstName LastName' },
@@ -63,4 +63,4 @@ export const MarketingTeamPage = () => {
   )
 }
 
-export default MarketingTeamPage
+export default EventTeamPage
